src/main/java/com/course_project/airlines/services/CustomUserDetailsService.java
package com.course_project.airlines.services;

import com.course_project.airlines.repositories.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CustomUserDetailsService implements UserDetailsService {
    private final UserRepository userRepository;

    /**
     * Method that loads the user's data entity by their email address
     * @param email - email address of the current user
     * @return - entity containing the user's details
     * @throws UsernameNotFoundException - if no user is found with the given email address
     */
    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        return userRepository.findByEmail(email);
    }
}
